Updated the GS 1826-24 data file and get_obs to read correctly
Previous version of the data file mixed up some of the values and also omitted the burst recurrence time. Corrected values and added the recurrence time as additional columns.

Extensive modification of get_obs to simplify, perform some checks on the input, and assemble the data correctly for "ensemble" mode runs; that is, replacing the burst times with the recurrence times, and keeping in all the bursts (cf. with the deletion of the "reference" bursts for the conventional mode)