Removed the 0.65 correction to yb (trec)
This correction was introduced to improve the agreement between settle and
Kepler, by Goodwin et al. (2019). The goal was to reduce the ignition
depth to better match the recurrence time and fluence; however because the
fluence is calculated from y and not yb, it seems this correction would
only be partially successful For later versions we make any corrections in
the Python code to avoid changes to the c code